+++ tests/Mocks/StaticCallable.php
@@ -4,13 +4,13 @@
  *
  * @link      https://github.com/slimphp/Slim
  * @copyright Copyright (c) 2011-2017 Josh Lockhart
- * @license   https://github.com/slimphp/Slim/blob/3.x/LICENSE.md (MIT License)
+ * @license   https://github.com/slimphp/Slim/blob/4.x/LICENSE.md (MIT License)
  */
 namespace Slim\Tests\Mocks;
 
 /**
-* Mock object for Slim\Tests\RouteTest
-*/
+ * Mock object for Slim\Tests\RouteTest
+ */
 class StaticCallable
 {
     public static function run($req, $res, $next)