+++ tests/RouterTest.php
@@ -4,7 +4,7 @@
  *
  * @link      https://github.com/slimphp/Slim
  * @copyright Copyright (c) 2011-2017 Josh Lockhart
- * @license   https://github.com/slimphp/Slim/blob/3.x/LICENSE.md (MIT License)
+ * @license   https://github.com/slimphp/Slim/blob/4.x/LICENSE.md (MIT License)
  */
 namespace Slim\Tests;
 
@@ -66,7 +66,6 @@ class RouterTest extends TestCase
 
     /**
      * Base path is ignored by relativePathFor()
-     *
      */
     public function testRelativePathFor()
     {